Revert "Kenta bed unity"
