--- GameCreatorTGS2020/Assets/Resources/output_images.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 180c60426fcac304fa723330eefdb18d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
