+++ GameCreatorTGS2020/Assets/Scenes/Replay.unity
@@ -293,7 +293,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &936093895
 RectTransform:
   m_ObjectHideFlags: 0
@@ -454,75 +454,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1067203256}
   m_CullTransparentMesh: 0
---- !u!1001 &1077116837
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 3449133221371723087, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_Name
-      value: New Sprite1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 4
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3449133221371723085, guid: a5e57bcc4f14cef4a94eaf175bed6cb7,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: a5e57bcc4f14cef4a94eaf175bed6cb7, type: 3}
 --- !u!1 &1134550476
 GameObject:
   m_ObjectHideFlags: 0
@@ -534,6 +465,7 @@ GameObject:
   - component: {fileID: 1134550480}
   - component: {fileID: 1134550479}
   - component: {fileID: 1134550478}
+  - component: {fileID: 1134550477}
   m_Layer: 0
   m_Name: EventSystem
   m_TagString: Untagged
@@ -541,6 +473,20 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
+--- !u!114 &1134550477
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1134550476}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4da27aac43b16714a8b82bbbd7133ecc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  rawImage: {fileID: 455934248}
+  Image: {fileID: 455934247}
 --- !u!114 &1134550478
 MonoBehaviour:
   m_ObjectHideFlags: 0
