--- GameCreatorTGS2020/Assets/Script/Maine/replay/GameObject.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 24f42bb99ee7814489291e44aec3d9e0
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
