--- GameCreatorTGS2020/Assets/Script/Maine/replay/MoveTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 107e2840f99f2de40ab2ad4cbad36220
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
