--- GameCreatorTGS2020/Assets/Script/Maine/replay/MoveTest1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c357e2328c1aee94da0c3a42e0abe8f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
