--- GameCreatorTGS2020/Assets/Script/Maine/replay/New Sprite.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3e983effe29de384eb48641217470d4e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
