--- GameCreatorTGS2020/Assets/Script/Maine/replay/New Sprite1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a5e57bcc4f14cef4a94eaf175bed6cb7
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
