--- GameCreatorTGS2020/Assets/Script/Maine/replay/WCC.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f30f0b79c3342c4a9fac6ffe1575a56
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
