--- GameCreatorTGS2020/Assets/Script/Maine/replay/WebCamStoper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3560334cb184a34594ba94a8d8eaebe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
