+++ GameCreatorTGS2020/Assets/Script/Maine/replay/WebCameraController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f164b4ebd600b6d479773fa73f7290e8
+guid: 458eb22f0325253479e9614bef7d5c2d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
