--- GameCreatorTGS2020/Assets/Script/Maine/replay/WebCameraController1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1d2d1618d49c0b64e8225c61d829c749
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
