--- GameCreatorTGS2020/Assets/Script/Maine/replay/WebCameraController2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 458eb22f0325253479e9614bef7d5c2d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
