--- GameCreatorTGS2020/Assets/Script/Ranking/outinput_image.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b17fc124a801c3f4da0958b2647208de
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
